+++ reactive-spring-session-mysql/src/test/java/net/andreaskluth/session/mysql/ReactiveMySQLSessionRepositoryMetricsTest.java
@@ -12,6 +12,7 @@ import java.time.Clock;
 import java.time.ZoneId;
 import java.util.Set;
 import net.andreaskluth.session.core.ReactiveVertxSessionRepository;
+import net.andreaskluth.session.core.ReactiveVertxSessionRepository.ReactiveSession;
 import net.andreaskluth.session.core.serializer.JdkSerializationStrategy;
 import net.andreaskluth.session.mysql.testsupport.MySQLDbExtension;
 import org.junit.jupiter.api.AfterEach;
@@ -45,8 +46,8 @@ class ReactiveMySQLSessionRepositoryMetricsTest {
     SimpleMeterRegistry simple = new SimpleMeterRegistry(SimpleConfig.DEFAULT, new MockClock());
     Metrics.addRegistry(simple);
 
-    var repo = sessionRepository();
-    var session = repo.createSession().block();
+    ReactiveVertxSessionRepository repo = sessionRepository();
+    ReactiveSession session = repo.createSession().block();
     repo.save(session).block();
     repo.findById(session.getId()).block();
     repo.cleanupExpiredSessions().block();
@@ -72,7 +73,7 @@ class ReactiveMySQLSessionRepositoryMetricsTest {
   }
 
   private ReactiveVertxSessionRepository sessionRepository() {
-    var sessionRepository =
+    ReactiveVertxSessionRepository sessionRepository =
         new ReactiveVertxSessionRepository(
             pool,
             new ReactiveMySQLSessionRepositoryQueries(),