+++ reactive-spring-session-postgres/src/test/java/net/andreaskluth/session/postgres/ReactivePostgresSessionRepositoryTest.java
@@ -5,8 +5,10 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import com.opentable.db.postgres.junit5.EmbeddedPostgresExtension;
 import com.opentable.db.postgres.junit5.PreparedDbExtension;
+import io.vertx.core.buffer.Buffer;
 import io.vertx.pgclient.PgException;
 import io.vertx.sqlclient.Pool;
+import io.vertx.sqlclient.Tuple;
 import java.io.Serializable;
 import java.lang.reflect.Field;
 import java.sql.Connection;
@@ -15,8 +17,10 @@ import java.time.Duration;
 import java.time.Instant;
 import java.time.ZoneId;
 import java.util.Objects;
+import java.util.concurrent.ExecutionException;
 import net.andreaskluth.session.core.ReactiveVertxSessionRepository;
 import net.andreaskluth.session.core.ReactiveVertxSessionRepository.ReactiveSession;
+import net.andreaskluth.session.core.serializer.DeserializationException;
 import net.andreaskluth.session.core.serializer.JdkSerializationStrategy;
 import net.andreaskluth.session.core.serializer.SerializationException;
 import net.andreaskluth.session.core.support.ReactiveSessionSchemaPopulator;
@@ -55,16 +59,16 @@ class ReactivePostgresSessionRepositoryTest {
 
   @Test
   void saveAndLoadWithAttributes() {
-    var repo = sessionRepository();
-    var session = repo.createSession().block();
+    ReactiveVertxSessionRepository repo = sessionRepository();
+    ReactiveSession session = repo.createSession().block();
 
     session.setAttribute(KEY, VALUE);
 
-    var sessionId = session.getId();
+    String sessionId = session.getId();
 
     repo.save(session).block();
 
-    var loadedSession = repo.findById(sessionId).block();
+    ReactiveSession loadedSession = repo.findById(sessionId).block();
 
     assertThat(loadedSession.getId()).isEqualTo(sessionId);
     assertThat(loadedSession.<String>getAttribute(KEY)).isEqualTo(VALUE);
@@ -72,11 +76,11 @@ class ReactivePostgresSessionRepositoryTest {
 
   @Test
   void duplicateSessionIdsAreNotPermitted() {
-    var repo = sessionRepository();
-    var session = repo.createSession().block();
+    ReactiveVertxSessionRepository repo = sessionRepository();
+    ReactiveSession session = repo.createSession().block();
     repo.save(session).block();
 
-    var anotherSession = repo.createSession().block();
+    ReactiveSession anotherSession = repo.createSession().block();
 
     setSessionId(anotherSession, session.getId());
 
@@ -85,19 +89,19 @@ class ReactivePostgresSessionRepositoryTest {
 
   @Test
   void saveAndLoadRemovingAttributes() {
-    var repo = sessionRepository();
+    ReactiveVertxSessionRepository repo = sessionRepository();
 
-    var session = repo.createSession().block();
+    ReactiveSession session = repo.createSession().block();
     session.setAttribute(KEY, VALUE);
 
     repo.save(session).block();
 
-    var loadedSession = repo.findById(session.getId()).block();
+    ReactiveSession loadedSession = repo.findById(session.getId()).block();
     loadedSession.removeAttribute(KEY);
 
     repo.save(loadedSession).block();
 
-    var reloadedSession = repo.findById(session.getId()).block();
+    ReactiveSession reloadedSession = repo.findById(session.getId()).block();
 
     assertThat(reloadedSession.getId()).isEqualTo(session.getId());
     assertThat(loadedSession.<String>getAttribute(KEY)).isNull();
@@ -105,41 +109,41 @@ class ReactivePostgresSessionRepositoryTest {
 
   @Test
   void findUnknownSessionIdShouldReturnNull() {
-    var repo = sessionRepository();
+    ReactiveVertxSessionRepository repo = sessionRepository();
     ReactiveSession session = repo.findById("unknown").block();
     assertThat(session).isNull();
   }
 
   @Test
   void deleteSessionById() {
-    var repo = sessionRepository();
-    var session = repo.createSession().block();
+    ReactiveVertxSessionRepository repo = sessionRepository();
+    ReactiveSession session = repo.createSession().block();
     repo.save(session).block();
 
     repo.deleteById(session.getId()).block();
 
-    var loadedSession = repo.findById(session.getId()).block();
+    ReactiveSession loadedSession = repo.findById(session.getId()).block();
     assertThat(loadedSession).isNull();
   }
 
   @Test
   void deleteSessionByIdWithUnknownSessionIdShouldNotCauseAnError() {
-    var repo = sessionRepository();
+    ReactiveVertxSessionRepository repo = sessionRepository();
     repo.deleteById("unknown").block();
   }
 
   @Test
   void rotateSessionIdChangesSessionId() {
-    var repo = sessionRepository();
-    var session = repo.createSession().block();
+    ReactiveVertxSessionRepository repo = sessionRepository();
+    ReactiveSession session = repo.createSession().block();
     session.setAttribute(KEY, VALUE);
 
-    var sessionId = session.getId();
+    String sessionId = session.getId();
     String changedSessionId = session.changeSessionId();
 
     repo.save(session).block();
 
-    var loadedSession = repo.findById(changedSessionId).block();
+    ReactiveSession loadedSession = repo.findById(changedSessionId).block();
 
     assertThat(sessionId).isNotEqualTo(changedSessionId);
     assertThat(loadedSession.getId()).isEqualTo(changedSessionId);
@@ -148,14 +152,14 @@ class ReactivePostgresSessionRepositoryTest {
 
   @Test
   void updatingValuesInSession() {
-    var repo = sessionRepository();
-    var session = repo.createSession().block();
+    ReactiveVertxSessionRepository repo = sessionRepository();
+    ReactiveSession session = repo.createSession().block();
 
     session.setAttribute(KEY, VALUE);
 
     repo.save(session).block();
 
-    var reloadedSession = repo.findById(session.getId()).block();
+    ReactiveSession reloadedSession = repo.findById(session.getId()).block();
 
     reloadedSession.setAttribute(KEY, "another value");
 
@@ -166,8 +170,8 @@ class ReactivePostgresSessionRepositoryTest {
 
   @Test
   void updatingWithSameValueShouldChangeSession() {
-    var repo = sessionRepository();
-    var session = repo.createSession().block();
+    ReactiveVertxSessionRepository repo = sessionRepository();
+    ReactiveSession session = repo.createSession().block();
 
     session.setAttribute(KEY, VALUE);
     session.clearChangeFlags();
@@ -178,8 +182,8 @@ class ReactivePostgresSessionRepositoryTest {
 
   @Test
   void addingNullValueForNewKeyShouldChangeSession() {
-    var repo = sessionRepository();
-    var session = repo.createSession().block();
+    ReactiveVertxSessionRepository repo = sessionRepository();
+    ReactiveSession session = repo.createSession().block();
 
     session.clearChangeFlags();
     session.setAttribute(KEY, null);
@@ -189,22 +193,22 @@ class ReactivePostgresSessionRepositoryTest {
 
   @Test
   void storeComplexObjectsInSession() {
-    var repo = sessionRepository();
-    var session = repo.createSession().block();
+    ReactiveVertxSessionRepository repo = sessionRepository();
+    ReactiveSession session = repo.createSession().block();
 
     session.setAttribute(KEY, new Complex(Instant.MAX));
 
     repo.save(session).block();
 
-    var reloadedSession = repo.findById(session.getId()).block();
+    ReactiveSession reloadedSession = repo.findById(session.getId()).block();
 
     assertThat(reloadedSession.<Complex>getAttribute(KEY)).isEqualTo(new Complex(Instant.MAX));
   }
 
   @Test
   void objectsThatAreNotSerializableShouldRaise() {
-    var repo = sessionRepository();
-    var session = repo.createSession().block();
+    ReactiveVertxSessionRepository repo = sessionRepository();
+    ReactiveSession session = repo.createSession().block();
 
     session.setAttribute(KEY, new NotSerializable(Instant.MAX));
 
@@ -213,9 +217,38 @@ class ReactivePostgresSessionRepositoryTest {
   }
 
   @Test
+  void objectsThatAreNotDeserializableShouldRaise()
+      throws ExecutionException, InterruptedException {
+    ReactiveVertxSessionRepository repo = sessionRepository();
+
+    ReactiveSession session = repo.createSession().block();
+    repo.save(session).block();
+
+    messWithSessionData(session);
+
+    Mono<ReactiveSession> reloadedSession = repo.findById(session.getId());
+    assertThatThrownBy(reloadedSession::block).isInstanceOf(DeserializationException.class);
+  }
+
+  @Test
+  void objectsThatAreNotDeserializableShouldNotRaiseIfConfigured()
+      throws ExecutionException, InterruptedException {
+    ReactiveVertxSessionRepository repo = sessionRepository();
+    repo.setInvalidateSessionOnDeserializationError(true);
+
+    ReactiveSession session = repo.createSession().block();
+    repo.save(session).block();
+
+    messWithSessionData(session);
+
+    ReactiveSession reloadedSession = repo.findById(session.getId()).block();
+    assertThat(reloadedSession).isNull();
+  }
+
+  @Test
   void savingMultipleTimes() {
-    var repo = sessionRepository();
-    var session = repo.createSession().block();
+    ReactiveVertxSessionRepository repo = sessionRepository();
+    ReactiveSession session = repo.createSession().block();
 
     session.setAttribute("keyA", "value A");
     Mono<Void> saveA = repo.save(session);
@@ -225,48 +258,48 @@ class ReactivePostgresSessionRepositoryTest {
     Mono<Void> saveB = repo.save(session);
     saveB.block();
 
-    var reloadedSession = repo.findById(session.getId()).block();
+    ReactiveSession reloadedSession = repo.findById(session.getId()).block();
     assertThat(reloadedSession.<String>getAttribute("keyA")).isEqualTo("value A");
     assertThat(reloadedSession.<String>getAttribute("keyB")).isEqualTo("value B");
   }
 
   @Test
   void savingInParallel() {
-    var repo = sessionRepository();
-    var session = repo.createSession().block();
+    ReactiveVertxSessionRepository repo = sessionRepository();
+    ReactiveSession session = repo.createSession().block();
 
     session.setAttribute("keyA", "value A");
-    var saveA = repo.save(session);
+    Mono<Void> saveA = repo.save(session);
 
     session.setAttribute("keyB", "value B");
-    var saveB = repo.save(session);
+    Mono<Void> saveB = repo.save(session);
 
     Flux.concat(saveA, saveB).blockLast();
 
-    var reloadedSession = repo.findById(session.getId()).block();
+    ReactiveSession reloadedSession = repo.findById(session.getId()).block();
     assertThat(reloadedSession.<String>getAttribute("keyA")).isEqualTo("value A");
     assertThat(reloadedSession.<String>getAttribute("keyB")).isEqualTo("value B");
   }
 
   @Test
   void expiredSessionsCanNotBeRetrieved() {
-    var repo = sessionRepository();
+    ReactiveVertxSessionRepository repo = sessionRepository();
     repo.setDefaultMaxInactiveInterval(Duration.ZERO);
 
-    var session = repo.createSession().block();
+    ReactiveSession session = repo.createSession().block();
     repo.save(session).block();
     assertThat(session.isExpired()).isTrue();
 
-    var loadedSession = repo.findById(session.getId()).block();
+    ReactiveSession loadedSession = repo.findById(session.getId()).block();
     assertThat(loadedSession).isNull();
   }
 
   @Test
   void expiredSessionsArePurgedByCleanup() {
-    var repo = sessionRepository();
+    ReactiveVertxSessionRepository repo = sessionRepository();
     repo.setDefaultMaxInactiveInterval(Duration.ZERO);
 
-    var session = repo.createSession().block();
+    ReactiveSession session = repo.createSession().block();
     repo.save(session).block();
 
     Integer count = repo.cleanupExpiredSessions().block();
@@ -290,6 +323,15 @@ class ReactivePostgresSessionRepositoryTest {
     Field sessionIdField = ReflectionUtils.findField(ReactiveSession.class, "sessionId");
     ReflectionUtils.makeAccessible(sessionIdField);
     ReflectionUtils.setField(sessionIdField, anotherSession, sessionId);
+  }
+
+  private void messWithSessionData(ReactiveSession session)
+      throws ExecutionException, InterruptedException {
+    pool.preparedQuery("UPDATE session SET session_data = $1 WHERE session_id = $2;")
+        .execute(Tuple.of(Buffer.buffer(new byte[] {1, 2, 3, 4, 5, 6}), session.getId()))
+        .toCompletionStage()
+        .toCompletableFuture()
+        .get();
   }
 
   @SuppressWarnings("UnusedVariable")